examples/proxy: add option to use send/recvmsg instead of send/recv
Generally not a good idea if you don't need it, as it has extra
overhead compared to the non-msg variants. Particularly for this use
case as we don't have a need for it, and we just put the iov and msghdr
on the stack and hence need to submit before leaving that scope. That
could obviously be avoided, but don't want to bother with that.

This will break with SQPOLL as we don't know when it has consumed them.
Hence that is disabled. If the lifetime issue got fixed, then we could
remove both the extra submits and enable it with SQPOLL.

Signed-off-by: Jens Axboe <axboe@kernel.dk>