examples/proxy: add support for provided buffers for sends
If provided buffers are supported by the kernel for send/sendmsg, then
we don't need to serialize our sends. This reduces the complexity of
dealing with send, which otherwise need to be serialized and hence proxy
has a backlog of sends to process if we end up with more than 1 inflight.

This can be turned on with the -u switch. By default it'll get enabled
if the kernel supports it, and any use of -u will override that. If set
to 1 and the kernel doesn't support it, we'll throw a warning and turn
it off.

Signed-off-by: Jens Axboe <axboe@kernel.dk>