examples/send-zerocopy: set CPU in io-wq mask
The normal affinity helper sets it appropriately, but the io-wq side
does not. Should not really matter as no io-wq activity should be
seen, honestly not even sure why the example a) sets io-wq affinity
at all, and b) not then have a way to set it to something close to
the other CPU, but not the same.

Fixes: https://github.com/axboe/liburing/issues/1070
Signed-off-by: Jens Axboe <axboe@kernel.dk>