examples/proxy: don't set iov->iov_base for ring provided sends
Not really important for how it works, as it just gets ignored by the
kernel as it needs to pick a buffer. But it's important in terms of
illustrating that userspace need not set it. Only the length can be
important here, as the send length will be the minimum of the passed
in length and the length in the picked buffer.

Signed-off-by: Jens Axboe <axboe@kernel.dk>