examples/proxy: wire up multishot mode for send/sendmsg
If we use multishot, then we only need to arm a single send per
data direction for each event loop we're doing. As long as the buffers
are added to the appropriate group, that's all the app needs to do to
ensure that all pending sends will go out next time submission is done.

Signed-off-by: Jens Axboe <axboe@kernel.dk>