Fix off-by-one in minor check for io_uring_check_version()
Ditto for the compile time version, IO_URING_CHECK_VERSION(). If
the majors are the same, only if the major is larger should we return
true.

Both of them really work the opposite way of what you'd expect, at
least logically. But better to leave them alone at this point and just
ensure the documentation is correct.

Signed-off-by: Jens Axboe <axboe@kernel.dk>