Introduce make uninstall in Makefile
Previously there was no uninstall feature, which was not user-friendly.
So now this commit adds the uninstall feature to the Makefile.
Following the installation funciton, we add two uninstallations target:
  1-'uninstall': for src dir uninstallation
  2-'uninstall-tests': for test dir uninstallation

In order to make sure that the path variables are valid, we also
include the 'config-host.mak' when performing the uninstall function.

Signed-off-by: Xinghui Li <korantli@tencent.com>