examples/proxy: size the CQ ring more reasonably
Should probably be an option and really mostly depends on the number
of buffers and the number of connections we have. For a single
connection, the nr_bufs size would be fine. For many connections, since
we drive everything with a single ring in this example, it should be
big enough to not overflow.

Should add overflow detection, just so we can warn that the performance
is then wore than it could be.

Signed-off-by: Jens Axboe <axboe@kernel.dk>