examples/proxy: add option to use huge pages for the ring
All the app needs to do is set IORING_SETUP_NO_MMAP to make this happen,
and liburing will try and use a huge page for the ring.

This is wasteful for small rings obviously, and if we use huge pages,
we'd also want the send/recv buffer rings to use them as well. That
will come in a later patch.

Signed-off-by: Jens Axboe <axboe@kernel.dk>