Change buffer ring head return value to uint16_t
We currently return an unsigned int from:

io_uring_buf_ring_head()
io_uring_buf_ring_available()

which is confusing since the head itself is an uint16_t. This makes it
easy to have a bug in an application using this, as it needs to cast
this value to uint16_t before doing math with the application known
buffer ring tail.

Since we haven't released a liburing version with this yet, clean it
up and use uint16_t consistently.

Signed-off-by: Jens Axboe <axboe@kernel.dk>