man/io_uring_buf_ring_add: add a note on buffer ID <-> ptr mapping
As long as the application knows how to map a given buffer ID to a
suitable pointer, then it can use whatever mappings it wants for
buffers. There's no liburing or kernel assumption that these mappings
are persistent over time, they can very well be different every time
a given buffer ID is added to the provided buffer ring.

Signed-off-by: Jens Axboe <axboe@kernel.dk>