test/send_recv: set IORING_SETUP_SUBMIT_ALL
The test_invalid() relies on both being submitted even though it
encounters an error at prep time, and on newer kernels with the
improved async setup, this means it'll only submit 1 of the 2
requests.

Set IORING_SETUP_SUBMIT_ALL to tell the kernel to keep submitting,
even in case of errors.

Signed-off-by: Jens Axboe <axboe@kernel.dk>