t/socket-io-cmd: do single retry for SIOCINQ
Sometimes we find legitimate data in the socket, which is a bit of a
mystery, but it should not fail the test case as the data is indeed
there and it isn't a test failure. It just arrives between the ioctl
and io_uring SIOCINQ.

Add a single retry in case this happens, as it should not fail the
test case.

Link: https://github.com/axboe/liburing/issues/1136
Signed-off-by: Jens Axboe <axboe@kernel.dk>