liburing-ffi.map: Add comments for symbols added in wrong version node
These symbols all appeared in the 2.6 version. One was added in 2.5,
but the symbol name in the map file was wrong so it was not exported.
The rest were all added in 2.6 but in the wrong nodes, either 2.4 or 2.5.

We cannot fix these now because the version is now attached to the
symbols, so the best we can do w/o breaking ABI is to document this
in the map file so that other people do not get confused.

Signed-off-by: Guillem Jover <guillem@hadrons.org>