test/accept-non-empty: add accept IORING_CQE_F_SOCK_NONEMPTY test
This tests whether we correctly get IORING_CQE_F_SOCK_NONEMPTY set
on an accept request, if there are more connections pending after
the current one has been accepted.

The test is currently gated on IORING_FEAT_RECVSEND_BUNDLE, as that
got added to the kernel at the same time as this feature. Not the
prettiest, but it's what we have for now...

Signed-off-by: Jens Axboe <axboe@kernel.dk>