configure: Respect relative prefix path
When the user passes a relative path, we end up splitting the
installation in multiple directories because it is relative to $CWD,
which changes when we recurse into subdirectories.

A common idiom I use is:

  ./configure --prefix=install ; make ; make install

and that currently results in part of the installation inside prefix
and part of it elsewhere:

 $ find . -type d  -name 'install'
  ./src/install
  ./install

Not biggy, but annoying.  Let's use the path where the configure command
was invoked as basedir, like other projects usually do.

Signed-off-by: Gabriel Krisman Bertazi <krisman@suse.de>
Link: https://lore.kernel.org/r/20240723232026.1444-1-krisman@suse.de
Signed-off-by: Jens Axboe <axboe@kernel.dk>