test/ooo-file-unreg: add out-of-order file update test
This was an issue years ago, but we never got around to adding a test
case for it. The kernel side fix back then was ensuring that updates
are processed in order, hence the test case is registering two files
and keeping an inflight request for file0, while updating first file1
and then file0.

Signed-off-by: Jens Axboe <axboe@kernel.dk>