test/recvsend_bundle: various fixes
This test was kind of broken, bring it more up to snuff and make it
resilient.

For some reason UDP doesn't seem to work reliably, unknown as of now
why that is. Test c generally hangs, with the sender side having sent
a bundle of 4 vecs, for 512b total. But the receiving side only gets
128b and even with a retry it gets -EAGAIN, indicating no data in the
socket. UDP stats do not show any dropped packets when this happens.
A mystery...

Signed-off-by: Jens Axboe <axboe@kernel.dk>