test/recvsend_bundle: enable UDP tests
The reason why UDP don't work reliably is that the bundles get in their
way. For example, if we bundle send 4 128b segments as one 512b segment,
then a 128b recv will just get 128b and a followup recv will not get
anything as that message is now done.

Just disable bundle testing for UDP.

Signed-off-by: Jens Axboe <axboe@kernel.dk>