man/io_uring_prep_timeout.3: correct IORING_TIMEOUT_ETIME_SUCCESS
The man page incorrectly states that the result code of a timed out
timer with IORING_TIMEOUT_ETIME_SUCCESS set will not post -ETIME as
the cqe->res result. But it has always done this, the flag only
affects how we treat dependent links on such a timer. If set, then
dependent links will not get failed, like they normally would on an
error condition.

Link: https://github.com/axboe/liburing/issues/1204
Signed-off-by: Jens Axboe <axboe@kernel.dk>