test/kallsyms: add test case reading /proc/kallsyms
The kernel is generally using fops->read_iter() for most things these
days, which is a great thing, however it means that we may not be
exercising the loop_rw_iter() path for reading/writing files depending
on how the test cases are configured. kallsyms is still using old
school style proc ->read(), so add a test case reading that file.

Signed-off-by: Jens Axboe <axboe@kernel.dk>