test/polfree: add a reasonable POLLFREE test case
The previous one was yanked because it was some syzbot mess that
only worked on specific archs. Write a "proper" one that leaves a
poll request pending (via a read) on a signalfd instance, that then
exits. The signal cleanup will invoke wake_up_pollfree(), which is
what we care about.

Loop it a bunch of times, and with the common ring flag types.

Signed-off-by: Jens Axboe <axboe@kernel.dk>