man/io_uring_enter: Clarify when EEXIST can occur
Clarify the restriction on what thread is allowed to issue the
io_uring_enter syscall for a given io_uring instance. If
IORING_GETEVENTS is not set no restrictions apply. If it is set the
submitting thread must be the thread that initially created the
io_uring instance in question or must not have the flag
IORING_SETUP_DEFER_TASKRUN set.

Signed-off-by: Constantin Pestka <constantin.pestka@c-pestka.de>