test: return skipped if we don't have permission to open file
If devices/filenames are passed in to the test case and we don't
have permission to open them, don't fail verbosely. Rather, just
skip the test case. This makes the tests less noisy when often run
as both root and as a normal user, allow you to use the same
configuration for both.

Signed-off-by: Jens Axboe <axboe@kernel.dk>