test/futex: don't leave futex wait pending
After a recent commit was merged that correctly frees the futex at
exit time, the futex test would randomly hang. This was debugged
as being because the alloc+free of the futex would get reused, and
race with a previous futex wait changing the memory that was now
being used for a new futex. This would prevent a wakeup from happening,
and hence the test would hang forever waiting for a completion.

In test_order(), before freeing the futex, issue a cancelation request
for the one futex wait that wasn't triggered. This ensures that it
won't get touched at cancelation time.

Signed-off-by: Jens Axboe <axboe@kernel.dk>