test: don't set IOSQE_IO_LINK on the last request
It's not valid to have LINK set across a submission boundary, hence
the last request being submitted will always get the link flag cleared
as it has nothing to link to.

This doesn't cause a kernel side issue, but it is misleading on the
test side, except for cases that deliberately check if the kernel
handles this condition.

Signed-off-by: Jens Axboe <axboe@kernel.dk>