test/runtests.sh: perform dmesg check as a normal user
For some reason DO_KMSG is limiting the dmesg checking as well, which
is both not necessary and useful as normal user. Skip the check. Yes
some distros disable dmesg checking for a normal user, but some do not.
And people running their own kernel can make sane choices. If the dmesg
call does fail, disable it from then on.

Signed-off-by: Jens Axboe <axboe@kernel.dk>