test/timeout: modify update timeout case for PREEMPT_RT
For some reason, on PREEMPT_RT, a just enqueued hrtimer will still
return that it's in the process of being executed and hence io_uring
will fail to extract it. Just overlook that case, by skipping the
elapsed time check if we ran into -EALREADY when attempting to
update the timer.

Signed-off-by: Jens Axboe <axboe@kernel.dk>