man/io_uring: Add remark about where to find info about memory ordering
The current man page has a remark about why certain barriers are
required when using the SQ and CQ and then goes on to say that memory
orderings (rightfully so) are out of scope, due to being a big topic.
This adds a pointer to the interested reader where to find this out
of scope information.

Signed-of-by: Constantin Pestka <constantin.pestka@c-pestka.de>