liburing: Pull load_acquire out of for loop to amortize cost
The io_uring_for_each_cqe macro is used to efficiently iterate over
all currently available CQEs. It allows to amortize the cost of the
store_release on the head for all CQEs processed via this macro.
This change also amortizes the load acquire on the tail, by pulling
the load out of the for loop.

Signed-off-by: Constantin Pestka <constantin.pestka@c-pestka.de>