Make io_uring_for_each_cqe() only read the tail once
There's no need to read the tail with acquire semantics for each loop,
it can just be done at the start of the loop. Iterating CQEs is
inherently racy anyway, if the ring is setup with signal notifications
for CQE postings.

Not the prettiest code in the world, but some hoops have to get jumped
through to avoid the compiler noticing that the passed in 'head' is now
effectively unused.

Signed-off-by: Jens Axboe <axboe@kernel.dk>