test/msg-ring: don't mistake T_EXIT_SKIP for failure
A recent commit botched this. Have the first part error if the test
had an error, or just return T_EXIT_SKIP if it got skipped. And fixup
the last case to just check for T_EXIT_FAIL before printing that the
test failed, when it fact it did not.

Fixes: bb01691f6b44 ("test/msg-ring: handle the skip case for older kernels")
Signed-off-by: Jens Axboe <axboe@kernel.dk>