examples/kdigest: remove "error" handling
There's on way the request will return -ECANCELED without being
explicitly canceled, which the example doesn't do, or the ring being
torn down. Just remove the attempted error handling here, as it's
incomplete/broken and it's better to have none (and just abort) than
broken error handling in what serves as an example.

Signed-off-by: Jens Axboe <axboe@kernel.dk>