lib: turn INT_FLAG_REG_RING into masked flag
If INT_FLAG_REG_RING matches IORING_ENTER_REGISTERED_RING, then it's
possible to drop the branches around masking in this value for every
io_uring_enter(2) system call.

This also allows for doing the same with future flags, where it would
be beneficial to have them set for each invocation, without needing
to add branches for that.

Signed-off-by: Jens Axboe <axboe@kernel.dk>