Add support for ring resizing
io_uring_resize_rings() support resizing the SQ and CQ rings, where
the most useful case here is resizing the CQ ring. This can be useful
for cases where users of a ring don't have a good idea of how busy it
will be. Rather than start with a big ring, they can start small for
the most common use cases, and enlarge it if needed if/when overflows
are seen.

Usage:

struct io_uring_params p = { };

p.sq_entries = new_sq_entries;
p.cq_entries = new_cq_entries;

io_uring_resize_rings(ring, &p);

where 'new_sq_entries' is the new value for the size of the SQ ring,
and 'new_cq_entries' is the new value for the size of the CQ ring.

Note that this is currently not support for rings setup with
IORING_SETUP_NO_MMAP.

Signed-off-by: Jens Axboe <axboe@kernel.dk>