test/rsrc_tags: use correct buffer index for test
The test is reading from buffer 0, but updating buffer 1. It's expecting
buffer 1 not to be put until the per-ring state allows it, but with
per-node refs, this is no longer true. However what is true, is that if
we use the right buffer, then the ordering will be observed.

Hence use buffer offset 1 for both the read and the update test, then
the test is actually testing what it should be testing. And it'll pass
without full buffer table pinning.

Signed-off-by: Jens Axboe <axboe@kernel.dk>