test/timeout: add timeout lock -> eventfd signal lock ordering
A recent kernel commit changed the timeout lock to be a raw spinlock,
but this can conflict with eventfd signaling which uses a regular lock.
This is usually fine as they are the same lock type, however on kernels
with PREEMPT_RT enabled, a raw spinlock is a classic spinlock, and the
normal spinlock is a sleeping lock. Nesting the sleeping lock inside
the raw spinlock will throw a warning.

Based on the syzbot test case.

Link: https://lore.kernel.org/io-uring/67710fe0.050a0220.226966.00bd.GAE@google.com/
Signed-off-by: Jens Axboe <axboe@kernel.dk>