test/read-inc-file: test incremental consumption with regular file
Incremental buffer usage on non-pollable files will commit the buffer
upfront. Nothing wrong with that, except the kernel had a bug where
it would commit the buffer and return the current buffer address.
After incrementing, the current address it the next read point, not
the current one. This caused the kernel to place the data later in
the userspace buffer than what the application expected.

Signed-off-by: Jens Axboe <axboe@kernel.dk>