Add Rust implementation of liburing.h
Update the code to function as a standalone crate that provides native
Rust implementations of all the `static inline` functions that live in
liburing.h.

GHA is updated to build on as many architectures as Rust supports. The
crate statically links in liburing.a and the liburing.h header is
updated to duplicate a declaration per `static inline` function
definition. This is done so the Rust code can reuse all of the existing
tests. If a new definition is added to liburing.h and the Rust
implementation lags, a linking error is generated when the
liburing_rs_test is built.

Signed-off-by: Christian Mazakas <christian.mazakas@gmail.com>