Hide io_uring_get_sqe symbol
The user-facing API for liburing has the user interface with
io_uring_get_sqe via a static inline function. However, this function
is also defined in the plain liburing archive, which can potential lead
to ODR violations.

We instead define it properly only for the liburing-ffi case and keep
it hidden when the liburing archive itself is built.

Signed-off-by: Christian Mazakas <christian.mazakas@gmail.com>