setup: deduplicate unmap steps
io_uring_queue_exit() repeats the code to unmap the SQE, SQ, and CQ
memory in 2 if cases. Simplify the if (A && B) else if (B) structure to
just if (B). Also prefer to use ring_entries instead of *kring_entries,
as kring_entries is deprecated and requires an additional indirection.

This was attempted before in commit d5049970cf12 ("setup: unify queue
exit path"), but this broke support for big SQEs, which only one of the
cases supported. That was fixed in commit 460c1b4ad781 ("src/setup: use
correct sqe size for unmapping").

Signed-off-by: Caleb Sander Mateos <csander@purestorage.com>