src/queue: IOPOLL does not need to enter the kernel for SQPOLL
A previous fix made the entering of the kernel for events reaping
conditional on IORING_SETUP_IOPOLL, but this isn't entirely correc.t
If SQPOLL is set, then the SQPOLL thread will do the work that the
application would have, if SQPOLL wasn't set.

Add an internal flag, INT_FLAG_CQ_ENTER, which is set for IOPOLL if
SQPOLL isn't also set. Then gate cq_ring_needs_enter() on this internal
flag, rather than IORING_SETUP_IOPOLL. That will ensure that the kernel
is only entered for the !SQPOLL case.

Reported-by: Matthias Jasny <matthiasjasny@gmail.com>
Fixes: 333561791386 ("Fix __io_uring_get_cqe() for IORING_SETUP_IOPOLL")
Signed-off-by: Jens Axboe <axboe@kernel.dk>