test/pipe-bug: ignore -ENOMEM on ring creation
Since this test case sets up and tears down a ton of rings, this can
result in the deferred freeing/unaccounting causing spurious -ENOMEM
returns. This is expected if the system is configured to be tight on
locked memory (ulimit -l). If this happens, inject a small delay and
just have it retry.

Link: https://github.com/axboe/liburing/issues/1377
Signed-off-by: Jens Axboe <axboe@kernel.dk>