Add support for IORING_OP_PIPE
This opcode creates a pipe, either with normal file descriptors or
with io_uring fixed descriptors. There are two helpers added for it,
which can aid in setting up an SQE for creating a pipe.

Add a test case as well that creates pipes with both normal and fixed
descriptors, and does a quick test (write to pipe, read from pipe, verify
data is good) as well.

Signed-off-by: Jens Axboe <axboe@kernel.dk>