Add __io_uring_clone_buffers() and __io_uring_clone_buffers_offset()
These variants don't implicitly set IORING_REGISTER_SRC_REGISTERED if
the source is registered, as that can be problematic if the calling
thread is different than the one that originally registered the ring
fd. Outside of that, no differences between the normal and __ variants
of the helpers, they accomplish the same thing.

Add an explanation of this in the man page as well.

Reported-by: Matthias Jasny <matthiasjasny@gmail.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>