test/sendmsg_iov_clean: fix stack use-after-free
Builds with address sanitizer seem to flag this test, noting an access
to the msghdr struct used in queue_sends(). This is because the address
of the stack-local is written to the SQE and placed in the SQ. Because
submission is not done in this stack frame, we wind up submitting
SQEs with addresses to long-dead stack-locals.

This commit relocates the msghdr struct along with the backing I/O
vectors to the same frame that does submission and asan becomes happy
with the test.

Signed-off-by: Christian Mazakas <christian.mazakas@gmail.com>