Add io_uring_memory_size() and io_uring_memory_size_params()
Apparently no helpers exists for gauging how much memory you need for
a ring with a given size and setup flags. io_uring_mlock_size() and
it's param friend do exist, but they will return 0 if
IORING_FEAT_NATIVE_WORKERS is set.

Add a set of helpers that can be used in conjunction with
io_uring_queue_init_mem(), so that an application may appropriately
allocate the required number of memory upfront for sharing between a
number of rings (or just for a single ring, whatever the need may be).

Signed-off-by: Jens Axboe <axboe@kernel.dk>