Add test case for short timerfd reads working
A recent kernel commit broke retry handling on anything backed by an
anonymous inode, such as eventfd (or timerfd, etc). Add a test case for
that.

Only regular files should wait on a full read. The test case will issue
a read for twice as much data being delivered compared to a single
timerfd trigger, however only a single trigger will happen. On a known
good kernel, the short read will be returned when the timerfd triggers.
On a broken kernel, io_uring will perform the read, and then decide to
wait for the latter half of the data. That never arrives, and the test
will abort after 1 second of waiting.

Signed-off-by: Jens Axboe <axboe@kernel.dk>