test/recv-mshot-fair: submit !DEFER receives at the same time
The test currently allows !DEFER to be unfair after resubmission,
because it would arm each recv as needed and submit it at the same time.
This means that once they are all done, R1 will be submitted separately
and process the remaining, and repeat for R2..R4. That causes
unfairness, as only one recv would be pending at the time for that case.

Keep track of pending submits and do them in one go, preserving fairness
even for !DEFER.

Signed-off-by: Jens Axboe <axboe@kernel.dk>