t/timestamp: Remove `#include <error.h>` and move `#include <time.h>` to the top
Remove `#include <error.h>` to fix musl libc build error:

  timestamp.c:2:10: fatal error: error.h: No such file or directory
      2 | #include <error.h>
        |          ^~~~~~~~~

And replace all usages of `error()` with the `t_error()` helper to
eliminate the dependency on <error.h>. Additionally, reorder the
includes to place <time.h> before <linux/errqueue.h> to address this
musl-specific error:

  In file included from timestamp.c:3:
  /usr/include/linux/errqueue.h:57:18: error: array type has incomplete element type 'struct timespec'
     57 |  struct timespec ts[3];
        |                  ^~

The reordering ensures that `struct timespec` is fully defined before
it's used in the kernel header. Not sure whether the include reordering
is the ideal-long-term-solution, but for now, at least it resolves the
current build failure on musl libc.

Fixes: 21224848af24d379d54fbf1bd43a60861fe19f9b ("tests: add a tx timestamp test")
Co-authored-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>