liburing.h: Add noexcept specifiers for C++
A new object macro `LIBURING_NOEXCEPT` is defined to `noexcept` when
compiling as C++ and left empty when compiling as C.  This macro is
applied to all exported and inline functions in the header files.  A
check for C++ >= 2011 is omitted because the library already assumes
this in other places.

Fixes: https://github.com/axboe/liburing/issues/1117
Signed-off-by: Rohan Hendrik Lean <rohan@rohanlean.de>