Add support for IORING_SETUP_CQE_MIXED
This adds core support for mixed sized CQEs in the same CQ ring.
Before this, CQEs were either 16b in size (the normal size), or 32b
if IORING_SETUP_CQE32 was set in the ring initialization. With the
mixed support, a CQE may be either 16b or 32b on the same CQ ring.
If the CQE is 32b in size, then IORING_CQE_F_32 will be set, and
the kernel may post a CQE with IORING_CQE_F_SKIP that should simply
be ignored as it's just a pad filler that is posted when required.

Signed-off-by: Jens Axboe <axboe@kernel.dk>