test/ringbuf-loop: add test case for incremental buf with overwrite
Add a test case that essentially overwrites the buffer meta data for
the buffer that is being used, with incremental consumption with
provided ring buffers. Some kernels had a bug where that could lead
to looping in the kernel, as the kernel was trying to commit an N byte
read but the buffer length got overwritten with zeroes. As there are
no other valid buffers in that ring, it would loop forever trying to
commit N bytes.

Based on a test case posted here:

https://lore.kernel.org/io-uring/tencent_2DEFFED135071FB225305A65D1688B303A09@qq.com/

and thanks to the Tencent guys for help with figuring this out and
providing a test case that showed the problem.

Signed-off-by: Jens Axboe <axboe@kernel.dk>