test/mock_file: use different message for permission denied
The kernel can either not have /dev/io_uring_mock if it hasn't been
enabled in the kernel, or if the kernel is too old. Or it can have
it, but the test program isn't run with sufficient credentials to
open the file. Use a separate warning message for those two
conditions.

Signed-off-by: Jens Axboe <axboe@kernel.dk>