man: include common -ENOMEM error for zero copy send
A common error to run into with zero copy send operations is that the
kernel needs to pin the memory for the duration of the send. The
amount pinned will be accounted as locked memory, hence the ulimit -l
setting needs to be large enough to accommodate the amount of memory
that is inflight. If it isn't, then the operation will be terminated
with -ENOMEM.

Signed-off-by: Jens Axboe <axboe@kernel.dk>