github: Test build against the installed liburing
When adding a new header file, the author might forget to add it to the
installation path in the Makefile.

For example, commit 7e565c0116ba ("tests: test the query interface"),
introduced an include to the query.h file via liburing.h, but the
query.h file was not added to the installation entry, which resulted
in the following error:

  In file included from a.c:1:
  /usr/include/liburing.h:19:10: fatal error: liburing/io_uring/query.h: No such file or directory
     19 | #include "liburing/io_uring/query.h"
        |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~
  compilation terminated.

This kind of breakage is only caught by the liburing that's installed
in the system. Make sure the GitHub CI covers it so we can catch the
breakage earlier before a new release is tagged and distributed.

Thanks to Romain Pereira for reporting the issue on GitHub!

Link: https://github.com/axboe/liburing/issues/1470
Cc: Romain Pereira <rpereira@anl.gov>
Cc: Pavel Begunkov <asml.silence@gmail.com>
Cc: Christian Mazakas <christian.mazakas@gmail.com>
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>