test: epwait: skip test_race on low CPU systems
The test_race hangs on systems with 2 or fewer CPUs when using
IORING_SETUP_SQPOLL, due to thread scheduling starvation.
Specifically, the main thread and SQPOLL thread occupy both CPUs,
leaving no CPU time for the writer thread to run, so on such systems
the test_race should be skipped.

Signed-off-by: Sebastian Chlad <sebastian.chlad@suse.com>