test/resize-rings: ensure T_EXIT_SKIP is properly passed back
The test case never returns T_EXIT_SKIP, as any -EINVAL will lead
to trying the DEFER case, which doesn't check for T_EXIT_SKIP and
return that. This can lead to missed failures of the test, where
a skip was done but it's returning that the test successfully
passed.

Signed-off-by: Jens Axboe <axboe@kernel.dk>