src/setup: unmap mmap'ed sqes pointer with the correct size
If allocating a huge page fails, buf_size will have already been
updated to a different size than the one the sqes pointer was
mapped with. Store the sqes mmap size in sqes_size, and use that
for unmapping.

This fixes a potential crash in applications, if no huge pages are
available and NO_MMAP is used for setup. Notably the reg-fd-only.c
and no-mmap-inval.c test cases were affected by that.

Signed-off-by: Jens Axboe <axboe@kernel.dk>