Add test case for updated POLL_ADD that then triggers
Add a test case for having a POLL_ADD request pending, then changing
the mask of it to something that then triggers. This tests both that
two CQEs are received, but also that the POLL_ADD result is what we
expect it to be. Both were an issue in older kernels.

Signed-off-by: Jens Axboe <axboe@kernel.dk>